backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
PyMuPDF
together
python-dotenv
huggingface-hub
